+++ tools/build.js
@@ -1,0 +1,9 @@
+const PipeTasks = require("./common/PipeTasks");
+PipeTasks(
+    [
+        {command: "build:ts"},
+        {command: "sync"}
+    ]
+).execute();
+
+