+++ tools/build_watch.js
@@ -1,0 +1,9 @@
+const PipeTasks = require("./common/PipeTasks");
+PipeTasks(
+    [
+        {command: "build:ts:watch"},
+        {command: "sync:watch"}
+    ]
+).execute();
+
+